chore(deps): update outdated deps
