+++ karma.conf.js
@@ -1,42 +1,36 @@
 // Karma configuration
 // Generated on Fri Jul 07 2017 15:50:08 GMT+0300 (MSK)
 
-module.exports = function(config) {
+module.exports = function (config) {
   config.set({
-
     // base path that will be used to resolve all patterns (eg. files, exclude)
     basePath: '',
-
 
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
     frameworks: ['karma-typescript', 'mocha', 'browserify'],
 
-
     // list of files / patterns to load in the browser
-    files: [
-      '**/*.ts'
-    ],
-
+    files: ['**/*.ts'],
 
     // list of files to exclude
     exclude: [
       'index.ts', // CLI-related tools should not be loaded into browser
-      'node_modules/**/*.ts'
+      'src/cli.ts',
+      'node_modules/**/*.ts',
     ],
-
 
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
     preprocessors: {
-      '**/*.ts': [ 'karma-typescript' ]
+      '**/*.ts': ['karma-typescript'],
     },
 
     typescriptPreprocessor: {
       tsconfigPath: './tsconfig.json',
-      ignorePath: function(path){ 
-       return /\.d\.ts$/.test(path);
-      }
+      ignorePath: function (path) {
+        return /\.d\.ts$/.test(path);
+      },
     },
 
     // test results reporter to use
@@ -44,28 +38,22 @@ module.exports = function(config) {
     // available reporters: https://npmjs.org/browse/keyword/karma-reporter
     reporters: ['progress'],
 
-
     // web server port
     port: 9876,
 
-
     // enable / disable colors in the output (reporters and logs)
     colors: true,
-
 
     // level of logging
     // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
     logLevel: config.LOG_INFO,
 
-
     // enable / disable watching file and executing tests whenever any file changes
     autoWatch: true,
-
 
     // start these browsers
     // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
     browsers: ['Firefox'],
-
 
     // Continuous Integration mode
     // if true, Karma captures browsers, runs the tests and exits
@@ -73,6 +61,6 @@ module.exports = function(config) {
 
     // Concurrency level
     // how many browser should be started simultaneous
-    concurrency: Infinity
-  })
-}
+    concurrency: Infinity,
+  });
+};
