+++ package.json
@@ -3,11 +3,14 @@
   "version": "1.0.7",
   "description": "i18n .po to .json file converter",
   "main": "dist/index.js",
+  "engines": {
+    "node": ">=12"
+  },
   "scripts": {
     "test": "karma start --browsers Firefox --single-run",
     "testwatch": "karma start --browsers Firefox",
-    "lint": "tslint index.ts 'src/**/*.ts' && prettier --check index.ts 'src/**/*.ts'",
-    "build": "tsc index.ts --outDir dist --declaration"
+    "lint": "eslint index.ts 'src/**/*.ts' && prettier --check index.ts 'src/**/*.ts'",
+    "build": "tsc index.ts --outDir dist --declaration && tsc src/cli.ts --outDir dist --declaration"
   },
   "keywords": [
     "i18n"
@@ -22,29 +25,33 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@types/mocha": "5.2.7",
-    "@types/node": "8.9.4",
-    "@types/yargs": "13.0.0",
-    "browserify": "16.3.0",
+    "@types/mocha": "10.0.1",
+    "@types/node": "18.11.18",
+    "@types/yargs": "17.0.20",
+    "@typescript-eslint/eslint-plugin": "5.49.0",
+    "@typescript-eslint/parser": "5.49.0",
+    "browserify": "17.0.0",
+    "eslint": "8.32.0",
+    "eslint-plugin-functional": "4.4.1",
     "i18n-proto": "1.0.5",
-    "karma": "4.2.0",
-    "karma-browserify": "6.1.0",
-    "karma-firefox-launcher": "1.1.0",
-    "karma-mocha": "1.3.0",
-    "karma-typescript": "4.1.1",
-    "mocha": "6.2.0",
-    "prettier": "1.18.2",
-    "ts-node": "8.3.0",
-    "tslib": "1.10.0",
-    "tslint": "5.18.0",
-    "tslint-consistent-codestyle": "1.15.1",
-    "tslint-eslint-rules": "5.4.0",
-    "tslint-microsoft-contrib": "6.2.0",
-    "typescript": "3.5.3",
-    "watchify": "3.11.1"
+    "karma": "6.4.1",
+    "karma-browserify": "8.1.0",
+    "karma-firefox-launcher": "2.1.2",
+    "karma-mocha": "2.0.1",
+    "karma-typescript": "5.5.3",
+    "mocha": "10.2.0",
+    "prettier": "2.8.3",
+    "ts-node": "10.9.1",
+    "tslib": "2.5.0",
+    "typescript": "4.9.4",
+    "watchify": "4.0.0"
   },
   "dependencies": {
-    "get-stdin": "7.0.0",
-    "yargs": "13.3.0"
+    "get-stdin": "8.0.0",
+    "yargs": "17.6.2"
+  },
+  "resolutions": {
+    "**/async": "3.2.2",
+    "karma-browserify/minimatch": "3.0.5"
   }
 }
