--- tslint.json
@@ -1,126 +1,0 @@
-{
-  "extends": ["tslint:recommended"],
-  "rulesDirectory": [
-    "node_modules/tslint-consistent-codestyle",
-    "node_modules/tslint-eslint-rules/dist/rules",
-    "node_modules/tslint-microsoft-contrib"
-  ],
-  "rules": {
-    "no-console": [false],
-    "no-namespace": [true, "allow-declarations"],
-    "no-null-keyword": true,
-    "only-arrow-functions": [true, "allow-declarations", "allow-named-functions"],
-    "object-literal-sort-keys": false,
-    "prefer-const": true,
-    "no-var-keyword": true,
-    "object-literal-shorthand": true,
-    "array-type": {
-      "options": ["array-simple"]
-    },
-    "prefer-array-literal": [
-      true,
-      {
-        "allow-type-parameters": true,
-        "allow-size-argument": true
-      }
-    ],
-    "interface-name": [true, "never-prefix"],
-    "no-eval": true,
-    "function-constructor": true,
-    "no-parameter-reassignment": true,
-    "ter-prefer-arrow-callback": [true],
-    "no-duplicate-imports": true,
-    "one-variable-per-declaration": [true, "ignore-for-loop"],
-    "increment-decrement": [true, "allow-post"],
-    "triple-equals": [true, "allow-null-check"],
-    "comment-format": [true, "check-space"],
-    "radix": true,
-    "no-construct": true,
-    "function-name": false,
-    "variable-name": [true, "check-format", "allow-leading-underscore", "allow-pascal-case"],
-    "no-this-assignment": true,
-    "import-name": false,
-    "no-string-literal": false,
-    "no-var-requires": false,
-    "valid-typeof": 2,
-    "ordered-imports": false,
-    "interface-over-type-literal": false,
-    "no-empty-interface": false,
-    "max-classes-per-file": false,
-    "ban-types": {
-      "options": [
-        ["Object", "Avoid using the `Object` type. Did you mean `object`?"],
-        ["Boolean", "Avoid using the `Boolean` type. Did you mean `boolean`?"],
-        ["Number", "Avoid using the `Number` type. Did you mean `number`?"],
-        ["String", "Avoid using the `String` type. Did you mean `string`?"],
-        ["Symbol", "Avoid using the `Symbol` type. Did you mean `symbol`?"]
-      ]
-    },
-    "ban": [
-      true,
-      {
-        "name": ["Object", "values"],
-        "message": "Object.values does not work in some browsers, use core/lib/object/values instead"
-      }
-    ],
-    "no-tautology-expression": true,
-    "no-useless-files": true,
-    // rules controlled by prettier https://github.com/prettier/tslint-config-prettier
-    "align": false,
-    "array-bracket-spacing": false,
-    "arrow-parens": false,
-    "block-spacing": false,
-    "brace-style": false,
-    "conditional-expression-parens": false,
-    "eofline": false,
-    "import-destructuring-spacing": false,
-    "import-spacing": false,
-    "indent": false,
-    "jsx-alignment": false,
-    "jsx-attribute-spacing": false,
-    "jsx-curly-spacing": false,
-    "jsx-equals-spacing": false,
-    "jsx-expression-spacing": false,
-    "jsx-no-closing-bracket-newline": false,
-    "jsx-no-multiline-js": false,
-    "jsx-space-before-trailing-slash": false,
-    "jsx-wrap-multiline": false,
-    "linebreak-style": false,
-    "literal-spacing": false,
-    "max-line-length": false,
-    "multiline-arrow": false,
-    "new-parens": false,
-    "newline-per-chained-call": false,
-    "no-consecutive-blank-lines": false,
-    "no-empty-line-after-opening-brace": false,
-    "no-extra-semi": false,
-    "no-irregular-whitespace": false,
-    "no-multi-spaces": false,
-    "no-semicolon-interface": false,
-    "no-trailing-whitespace": false,
-    "no-unnecessary-parens-for-arrow-function-arguments": false,
-    "no-unnecessary-semicolons": false,
-    "number-literal-format": false,
-    "object-curly-spacing": false,
-    "object-literal-key-quotes": false,
-    "one-line": false,
-    "quotemark": false,
-    "react-tsx-curly-spacing": false,
-    "semicolon": false,
-    "space-before-function-paren": false,
-    "space-in-parens": false,
-    "space-within-parens": false,
-    "ter-arrow-parens": false,
-    "ter-arrow-spacing": false,
-    "ter-computed-property-spacing": false,
-    "ter-func-call-spacing": false,
-    "ter-indent": false,
-    "ter-max-len": false,
-    "ter-no-irregular-whitespace": false,
-    "ter-no-tabs": false,
-    "ter-padded-blocks": false,
-    "trailing-comma": false,
-    "typedef-whitespace": false,
-    "whitespace": false
-  }
-}
